Support DECIMAL columns with precision-aware splitting

+++ v2/sourcedb-to-spanner/src/main/java/com/google/cloud/teleport/v2/source/reader/io/schema/SourceColumnIndexInfo.java
@@ -156,5 +156,5 @@ public abstract class SourceColumnIndexInfo implements Comparable<SourceColumnIn
           IndexType.BIG_INT_UNSIGNED, BigDecimal.class,
           IndexType.BINARY, BoundaryExtractorFactory.BYTE_ARRAY_CLASS,
           IndexType.TIME_STAMP, Timestamp.class,
-          IndexType.DECIMAL, BigInteger.class);
+          IndexType.DECIMAL, BigDecimal.class);
 }
